Development (#56)
* Update HOWTO.md

* bug

- silly bug when timestamp is all nan
- add an attribute that could be skipped

* Update HOWTO.md join trajectories

* correct bug with symbolic paths

* add different features param files for the WT2

* add abspath and realpath to avoid problems in Linux

This is a very bad bug where in linux the path resolution didn't understood the tempoerary and the destination file where the same. Therefore it sometimes delete the source files.

* replace debugging call with warning

* small bug due to using float as index

There were a couple of files where the timestamp_ind was `float`. I force it to be `int` and it fix the problem.

* bug

* bug