Features can now be calculates on a well-by-well basis
3 parameters added:
- MWP_total_n_wells (can be -1,  24, 48, 96). If -1, the FOV is not split
- MWP_whichsideup (can be upright or upside-down)
- MWP_well_shape (can be square or circle)

The FOV splitting happens in the FEAT_TIERPSY step of the analysis: an image is obtained by loading the /full_data from the masked videos and the FOV splitting is done based on that.
A table, /fov_wells, is added to the featuesN.hdf5 file, containing the following information:
- name of the well
- x_min, x_max, y_min, y_max in pixels for each well

The coordinates of the worms, taken from the /trajectories_data, are used to assign a well_name to each point of each trajectory. A "well_name" column is added to the /timeseries_data in featuresN.hdf5. This column contains either the well name or the string 'n/a' if either the worm was outside every recognised well or if the FOV splitting is not carried out (as chosen by the user via the parameter file)

If the analysis is carried on a well-by-well basis, the /features_stats will also have a well_name information. Instead of being a N_features-by-2 table with feature name and value as the only columns, the /feature_stats will now be a N_features*N_wells_in_FOV-by-3 table., with features names (will be repeated N_wells_in_FOV times), the corresponding values, and a third column with the well name each feature was measured in.