compressVideo.py: changed to take advantage of code Avelino had already introduced at a later stage
Avelino already allows subtracting the bgnd from the masked file, at the trajectory creation stage. The only condition is to have the flag is_full_bgnd_subtraction==True and (for it to be faster) to have a /bgnd field in the masked file. So I only put it as a default to save the /bgnd in the masked image file if the is_bgnd_subtraction==True