Development (#69)
* Update HOWTO.md

* bug

- silly bug when timestamp is all nan
- add an attribute that could be skipped

* Update HOWTO.md join trajectories

* correct bug with symbolic paths

* add different features param files for the WT2

* add abspath and realpath to avoid problems in Linux

This is a very bad bug where in linux the path resolution didn't understood the tempoerary and the destination file where the same. Therefore it sometimes delete the source files.

* replace debugging call with warning

* small bug due to using float as index

There were a couple of files where the timestamp_ind was `float`. I force it to be `int` and it fix the problem.

* bug

* bug

* add poses model and reorganize a bit

* add multiprocessing to summary calculation

* improve multiprocessing

* bug

* bug

again...

* fix bug related to opencv 4.0

* Fix matplotlib qt5 import

The import of qt5 as the backend of matplotlib should only occur on the GUI.
I removed the ploting functions of openworm since they were making an import of matplotlib before the gui was called.

* add MAGGOTS.json parameters

* remove limitation of having to run the SKEL_FILT step

* small modifications

* add MAGGOT parameters file

* comments